remove safeManagement and update Safe vue
